% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_evonet_misc.R
\name{initialize_evonet_misc}
\alias{initialize_evonet_misc}
\title{Title}
\usage{
initialize_evonet_misc(dat)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
