% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_infecteds_vl.R
\name{initialize_infecteds_vl}
\alias{initialize_infecteds_vl}
\title{Title}
\usage{
initialize_infecteds_vl(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
