% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_module.R
\name{initialize_module}
\alias{initialize_module}
\title{Title}
\usage{
initialize_module(x, param, init, control, s)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
