% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parameters_primary.R
\name{input_parameters_primary}
\alias{input_parameters_primary}
\title{Title}
\usage{
input_parameters_primary()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
