% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parameters_qa_qc.R
\name{input_parameters_qaqc}
\alias{input_parameters_qaqc}
\title{Title}
\usage{
input_parameters_qaqc(evonet_params)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
