% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_new_additions_both.R
\name{new_additions_fxn_both}
\alias{new_additions_fxn_both}
\title{Title}
\usage{
new_additions_fxn_both(input_list, dat, index, type = c("births", "initial"),
  at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
