% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_initial_diagnostic.R
\name{plots_initial_diagnostic}
\alias{plots_initial_diagnostic}
\title{Title}
\usage{
plots_initial_diagnostic(x, param, init, control, s = 1)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
