% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_popsumm.R
\name{plots_popsumm_ymin_ymax}
\alias{plots_popsumm_ymin_ymax}
\title{Title}
\usage{
plots_popsumm_ymin_ymax(variable, plot_type, popsumm)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
