% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_epimodel_control_object.R
\name{setup_epimodel_control_object}
\alias{setup_epimodel_control_object}
\title{Title}
\usage{
setup_epimodel_control_object(evonet_params, module_list)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
