% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_initialize_nw.R
\name{setup_initialize_network}
\alias{setup_initialize_network}
\title{Title}
\usage{
setup_initialize_network(params)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
