% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_condom_use.R
\name{social_condom_use}
\alias{social_condom_use}
\title{Title}
\usage{
social_condom_use(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
