% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_role_msm.R
\name{social_role_msm}
\alias{social_role_msm}
\title{Title}
\usage{
social_role_msm(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
