% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_popsumm.R
\name{summary_popsumm}
\alias{summary_popsumm}
\title{Title}
\usage{
summary_popsumm(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
