% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_deaths_module.R
\name{vital_deaths_module}
\alias{vital_deaths_module}
\title{Title}
\usage{
vital_deaths_module(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
